# Optimize recurrent_op using Prepare and RunPreparedContext, avoiding create operators in every iter.