`ExecutorPrepareContext` can be reused, and there is no need to create in every iters in `recurrent_op`. This PR is tested in PaddingRNN padding model.

|  model | before | after | speedup |
|---------|---------|------|-----------|
| small | 142.8 | 116.2 | 23% |
| large | 170.5 | 148.4 | 15% |